Pull syncGroup,syncContacts into own files
