+++ packages/shared/src/store/sync/logger.ts
@@ -1,0 +1,3 @@
+import { createDevLogger } from '../../debug';
+
+export const logger = createDevLogger('sync', false);
