+++ packages/shared/src/store/sync/updateLastActivityTime.ts
@@ -1,0 +1,6 @@
+import * as db from '../../db';
+
+// Update the last activity timestamp when we receive new data
+export function updateLastActivityTime() {
+  db.lastActivityAt.setValue(Date.now());
+}
